+++ lithops/serverless/backends/aws_lambda/config.py
@@ -45,7 +45,7 @@ RUNTIME_MEMORY_DEFAULT = 256  # Default memory: 256 MB
 RUNTIME_MEMORY_MAX = 10240  # Max. memory: 10240 MB
 
 MAX_CONCURRENT_WORKERS = 1000
-INVOKE_POOL_THREADS_DEFAULT = 500
+INVOKE_POOL_THREADS_DEFAULT = 64
 
 
 def load_config(config_data):
@@ -120,9 +120,6 @@ def load_config(config_data):
 
     if not all([efs_conf['mount_path'].startswith('/mnt') for efs_conf in config_data['aws_lambda']['efs']]):
         raise Exception("All mount paths must start with '/mnt' on 'aws_lambda/efs/*/mount_path' section")
-
-    if config_data['aws_lambda'].get('beta_features', False):
-        config_data['aws_lambda']['invoke_pool_threads'] = 64
 
     # Put credential keys to 'aws_lambda' dict entry
     config_data['aws_lambda'].update(config_data['aws'])