+++ lithops/storage/backends/aws_s3/aws_s3.py
@@ -30,16 +30,17 @@ class S3Backend:
     def __init__(self, s3_config):
         logger.debug("Creating S3 client")
         self.s3_config = s3_config
-        user_agent = s3_config['user_agent']
-        service_endpoint = s3_config['endpoint']
+        self.user_agent = s3_config['user_agent']
+        self.service_endpoint = s3_config['endpoint']
+        self.region = s3_config['region_name']
 
-        if 'http:' in service_endpoint:
+        if 'http:' in self.service_endpoint:
             logger.warning('Endpoint {} is insecure - it is recommended '
-                           'to change this to https://'.format(service_endpoint))
+                           'to change this to https://'.format(self.service_endpoint))
 
         client_config = botocore.client.Config(
             max_pool_connections=128,
-            user_agent_extra=user_agent,
+            user_agent_extra=self.user_agent,
             connect_timeout=CONN_READ_TIMEOUT,
             read_timeout=CONN_READ_TIMEOUT,
             retries={'max_attempts': OBJ_REQ_RETRIES}
@@ -49,11 +50,11 @@ class S3Backend:
             's3', aws_access_key_id=s3_config['access_key_id'],
             aws_secret_access_key=s3_config['secret_access_key'],
             config=client_config,
-            endpoint_url=service_endpoint
+            endpoint_url=self.service_endpoint
         )
 
         msg = STORAGE_CLI_MSG.format('S3')
-        logger.info("{} - Endpoint: {}".format(msg, service_endpoint))
+        logger.info("{} - Region: {}".format(msg, self.region))
 
     def get_client(self):
         '''