+++ lithops/executors.py
@@ -714,7 +714,3 @@ class StandaloneExecutor(FunctionExecutor):
                          workers=workers,
                          monitoring=monitoring,
                          log_level=log_level)
-
-    def create(self):
-        runtime_key, runtime_meta = self.compute_handler.create()
-        self.internal_storage.put_runtime_meta(runtime_key, runtime_meta)