+++ CHANGELOG.md
@@ -6,8 +6,10 @@
 - [Storage] Added MinIO storage backend
 - [Core] Allow to pass function args as part of the invocation payload in FaaS backends
 - [Core] Optimize call_async() calls with an internal function caching system
-- [AWS Lambda] New invoke method
-- [Core] Allow to set storage_bucket at storage backend level in config
+- [AWS Lambda] New invoke method that reduces total invocation time
+- [Core] Allow to set the bucket name at storage backend level in config
+- [localhost] stop running jobs processes on ctrl-c or exception
+- [AWS Lambda] region_name parameter in config
 
 ### Changed
 - [Core] Improved performance and efficiency of the lithops cleaner background process